continue refactoring, drop WeekMap usage
