+++ lib/SparceCoverageCollector.js
@@ -1,31 +1,7 @@
-const MaybeWeakMap = typeof WeakMap === 'function' ? WeakMap : function () {
-  const objectList = [];
-  const dataList = [];
-
-  this.set = function (object, data) {
-    const index = objectList.indexOf(object);
-    if (index !== -1) {
-      data[index] = data;
-    } else {
-      objectList.push(object);
-      dataList.push(data);
-    }
-  };
-
-  this.get = function (object) {
-    const index = objectList.indexOf(object);
-    if (index !== -1) {
-      return dataList[index];
-    }
-  };
-};
-
-
 class SparceCoverageCollector {
   constructor() {
     this.srcCoverage = {};
-    this.metaInfo = new MaybeWeakMap();
-
+    this.metaInfo = {};
   }
 
   getSourceCoverage(filename) {
@@ -40,19 +16,19 @@ class SparceCoverageCollector {
         b: {},
         f: {},
       };
-      this.metaInfo.set(data, {
+      this.metaInfo[filename] = {
         indexes: {},
         lastIndex: {
           s: 0,
           b: 0,
           f: 0,
         },
-      });
+      };
     }
 
     return {
       data,
-      meta: this.metaInfo.get(data),
+      meta: this.metaInfo[filename],
     };
   }
