+++ lib/remap.js
@@ -6,17 +6,15 @@ const path = require('path');
 const fs = require('fs');
 const { SourceMapConsumer } = require('source-map/lib/source-map-consumer');
 
-/* global WeakMap */
-
-const updateFunctions = require('./updateFunctions');
-const updateStatements = require('./updateStatements');
-const updateBranches = require('./updateBranches');
-
 const sourceMapRegEx = /(?:\/{2}[#@]{1,2}|\/\*)\s+sourceMappingURL\s*=\s*(data:(?:[^;]+;)+base64,)?(\S+)/;
 
 const SparceCoverageCollector = require('./SparceCoverageCollector');
 
 const getMapping = require('./getMapping');
+
+const remapFunction = require('./remapFunction');
+const remapBranch = require('./remapBranch');
+
 /**
  * Remaps coverage data based on the source maps it discovers in the
  * covered files and returns a coverage Collector that contains the remappped
@@ -185,9 +183,36 @@ function remap(coverage, options = {}) {
         return Object.assign(mapping, { source: resolvePath(mapping.source) });
       };
 
-      updateFunctions(fileCoverage, getMappingResolved, sparceCoverageCollector);
-      updateStatements(fileCoverage, getMappingResolved, sparceCoverageCollector);
-      updateBranches(fileCoverage, getMappingResolved, sparceCoverageCollector);
+      Object.keys(fileCoverage.branchMap).forEach((index) => {
+        const genItem = fileCoverage.branchMap[index];
+        const hits = fileCoverage.b[index];
+        const info = remapBranch(genItem, getMappingResolved);
+        if (info) {
+          sparceCoverageCollector.updateBranch(info.source, info.srcItem, hits);
+        }
+      });
+
+      Object.keys(fileCoverage.fnMap).forEach((index) => {
+        const genItem = fileCoverage.fnMap[index];
+        const hits = fileCoverage.f[index];
+
+        const info = remapFunction(genItem, getMappingResolved);
+
+        if (info) {
+          sparceCoverageCollector.updateFunction(info.source, info.srcItem, hits);
+        }
+      });
+
+      Object.keys(fileCoverage.statementMap).forEach((index) => {
+        const genItem = fileCoverage.statementMap[index];
+        const hits = fileCoverage.s[index];
+
+        const mapping = getMappingResolved(genItem);
+
+        if (mapping) {
+          sparceCoverageCollector.updateStatement(mapping.source, mapping.loc, hits);
+        }
+      });
 
       // todo: refactor exposing implementation details
       const srcCoverage = sparceCoverageCollector.getFilesCoverage();
